+++ src/components/CustomLink/CustomLink.css
@@ -1,0 +1,71 @@
+:root {
+  --min-custom-link-width: 72px;
+}
+
+.custom-link {
+  cursor: pointer;
+  font: var(--font-custom-link);
+  padding: 4px 12px;
+  border-radius: 4px;
+  text-decoration: none;
+  text-align: center;
+  display: inline-block;
+  border: 2px solid var(--color-custom-link-primary-initial-border);
+  white-space: nowrap;
+  min-width: var(--min-custom-link-width);
+  line-height: normal;
+}
+
+.custom-link_mode_primary {
+  background-color: var(--color-custom-link-primary-initial-background);
+  color: var(--color-custom-link-primary-initial-text);
+  border-color: var(--color-custom-link-primary-initial-border);
+}
+
+.custom-link_mode_primary:hover {
+  background-color: var(--color-custom-link-primary-hover-background);
+  color: var(--color-custom-link-primary-hover-text);
+  border-color: var(--color-custom-link-primary-hover-border);
+}
+
+.custom-link_mode_primary:active {
+  background-color: var(--color-custom-link-primary-active-background);
+  color: var(--color-custom-link-primary-active-text);
+  border-color: var(--color-custom-link-primary-active-border);
+}
+
+.custom-link_mode_secondary {
+  background-color: var(--color-custom-link-secondary-initial-background);
+  color: var(--color-custom-link-secondary-initial-text);
+  border-color: var(--color-custom-link-secondary-initial-border);
+}
+
+.custom-link_mode_secondary:hover {
+  background-color: var(--color-custom-link-secondary-hover-background);
+  color: var(--color-custom-link-secondary-hover-text);
+  border-color: var(--color-custom-link-secondary-hover-border);
+}
+
+.custom-link_mode_secondary:active {
+  background-color: var(--color-custom-link-secondary-active-background);
+  color: var(--color-custom-link-secondary-active-text);
+  border-color: var(--color-custom-link-secondary-active-border);
+}
+
+.custom-link_mode_dangerous {
+  background-color: var(--color-custom-link-dangerous-initial-background);
+  color: var(--color-custom-link-dangerous-initial-text);
+  border-color: var(--color-custom-link-dangerous-initial-border);
+}
+
+.custom-link_mode_dangerous:hover {
+  background-color: var(--color-custom-link-dangerous-hover-background);
+  color: var(--color-custom-link-dangerous-hover-text);
+  border-color: var(--color-custom-link-dangerous-hover-border);
+}
+
+.custom-link_mode_dangerous:active {
+  background-color: var(--color-custom-link-dangerous-active-background);
+  color: var(--color-custom-link-dangerous-active-text);
+  border-color: var(--color-custom-link-dangerous-active-border);
+}
