+++ src/components/CustomLink/CustomLink.tsx
@@ -1,12 +1,12 @@
 import block from "bem-cn";
-import "./Link.css";
+import "./CustomLink.css";
 import { Link, LinkProps } from "react-router-dom";
 
 interface CustomLinkProps extends LinkProps {
   mode?: "primary" | "secondary" | "dangerous";
 }
 
-const b = block("link");
+const b = block("custom-link");
 
 export default function CustomLink(props: CustomLinkProps) {
   const { mode = "secondary", className = "", to, children } = props;
